Added Initialization Photos
The start of the program now prompts the user to look at two dots and
press q while looking at them. These images are stored to the
refrence_images dictionary with the keys of 'downward' and 'neutral'.
The program now attempts to open in fullscreen windowed with a default
resolution of 1920x1080, which the user can change at the beginning of
the program. If no change is necessary just press 'enter' immediately.
If you can't find the program when it opens it is because it doesn't
open on top of other windows, check your taskbar. I'll fix the issue
later but I was having issue of the other window not being on top of the
fullscreened window.